# cmd/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing romantic events in SimpEngine project",
        "title": "SimpEngine API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/simp-target": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Views a SimpTargets of current User.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simp_target"
                ],
                "summary": "View SimpTargets by User",
                "responses": {
                    "200": {
                        "description": "List of SimpTargets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/simptarget.SimpTarget"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a simp target by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simp_target"
                ],
                "summary": "Creates simp target",
                "parameters": [
                    {
                        "description": "Simp target data",
                        "name": "simpTarget",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/simptarget.SimpTargetRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simptarget.CreateSimpTargetResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request (invalid ID, invalid body, or validation error)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid credentials)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/simp-target/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Views a SimpTarget by its ID. Users can only see their simp targets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simp_target"
                ],
                "summary": "View SimpTarget",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "SimpTarget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SimpTarget",
                        "schema": {
                            "$ref": "#/definitions/simptarget.SimpTarget"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing SimpTarget by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simp_target"
                ],
                "summary": "Update SimpTarget",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "SimpTarget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SimpTarget data",
                        "name": "simpTarget",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/simptarget.SimpTargetRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request (invalid ID, invalid body, or validation error)",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a SimpTarget by its ID. Only allowed if the user has proper permissions.",
                "tags": [
                    "simp_target"
                ],
                "summary": "Delete SimpTarget",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "SimpTarget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_BODY"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "simptarget.CreateSimpTargetResponseDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "simptarget.SimpTarget": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "simptarget.SimpTargetRequestDTO": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}